pyproject.toml
[project]
name = "fruitmaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.29.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]
